backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# PDF Processing
pdfplumber==0.11.4
pypdf==5.0.1

# Data Processing
pandas==2.2.3
pandas-stubs==2.2.3.241126
numpy==2.1.2

# Machine Learning
prophet==1.1.6
scikit-learn==1.5.2

# Export
reportlab==4.2.5
openpyxl==3.1.5

# Scheduler
APScheduler==3.10.4

# HTTP Client
httpx==0.27.2
PyJWT==2.9.0
passlib==1.7.4
argon2-cffi==23.1.0

# HTML Parsing
selectolax==0.4.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code Quality
black==24.10.0
ruff==0.7.1
mypy==1.13.0
types-passlib==1.7.7.20250602

# Utils
python-multipart==0.0.12
python-dateutil==2.9.0
